+++ src/adapter/bun/handler.ts
@@ -120,7 +120,18 @@ export const mapResponse = (
 					return handleStream(response as any, set, request) as any
 
 				if (typeof (response as Promise<unknown>)?.then === 'function')
-					return (response as Promise<unknown>).then((x) => mapResponse(x, set)) as any
+					return (response as Promise<unknown>).then((x) =>
+						mapResponse(x, set)
+					) as any
+
+				// custom class with an array-like value
+				// eg. Bun.sql`` result
+				if (Array.isArray(response))
+					return new Response(JSON.stringify(response), {
+						headers: {
+							'Content-Type': 'application/json'
+						}
+					}) as any
 
 				// @ts-expect-error
 				if (typeof response?.toResponse === 'function')
@@ -265,6 +276,15 @@ export const mapEarlyResponse = (
 				if (typeof response?.toResponse === 'function')
 					return mapEarlyResponse((response as any).toResponse(), set)
 
+				// custom class with an array-like value
+				// eg. Bun.sql`` result
+				if (Array.isArray(response))
+					return new Response(JSON.stringify(response), {
+						headers: {
+							'Content-Type': 'application/json'
+						}
+					}) as any
+
 				if ('charCodeAt' in (response as any)) {
 					const code = (response as any).charCodeAt(0)
 
@@ -382,6 +402,15 @@ export const mapEarlyResponse = (
 				// @ts-expect-error
 				if (typeof response?.toResponse === 'function')
 					return mapEarlyResponse((response as any).toResponse(), set)
+
+				// custom class with an array-like value
+				// eg. Bun.sql`` result
+				if (Array.isArray(response))
+					return new Response(JSON.stringify(response), {
+						headers: {
+							'Content-Type': 'application/json'
+						}
+					}) as any
 
 				if ('charCodeAt' in (response as any)) {
 					const code = (response as any).charCodeAt(0)
@@ -502,6 +531,15 @@ export const mapCompactResponse = (
 			// @ts-expect-errors
 			if (typeof response?.toResponse === 'function')
 				return mapCompactResponse((response as any).toResponse())
+
+			// custom class with an array-like value
+			// eg. Bun.sql`` result
+			if (Array.isArray(response))
+				return new Response(JSON.stringify(response), {
+					headers: {
+						'Content-Type': 'application/json'
+					}
+				}) as any
 
 			if ('charCodeAt' in (response as any)) {
 				const code = (response as any).charCodeAt(0)